+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added 
+- Publisher's site demo banner view
 
 ## [0.3.0] - 2018-10-02
 ### Changed
@@ -30,7 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - User email change
 - User password change and recovery    
 
-[Unreleased]: https://github.com/adshares/adserver/compare/v0.3.0...develop
+[Unreleased]: https://github.com/adshares/adserver/compare/v0.3.0...master
 [0.3.0]: https://github.com/adshares/adserver/compare/v0.2.1...v0.3.0
 [0.2.1]: https://github.com/adshares/adserver/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/adshares/adserver/compare/v0.1.0...v0.2.0